Fix UI bug when book conversion enabled on mobile
On mobile, when book conversion is enabled and using the cabliBlur theme, the upload menu appears behind the description text area making it difficult (though not impossible with some careful guesswork around where the buttons should be).

This commit fixes the styling of this section so that it no longer appears behind the description text area.